fix the upgrade command to heed the -n option in `remove'
modified: scripts/install-cm3-compiler.sh

+++ scripts/install-cm3-compiler.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-# $Id: install-cm3-compiler.sh,v 1.2 2003-03-21 16:14:42 wagner Exp $
+# $Id: install-cm3-compiler.sh,v 1.3 2003-07-19 21:23:00 wagner Exp $
 
 if [ -n "$ROOT" -a -d "$ROOT" ] ; then
   sysinfo="$ROOT/scripts/sysinfo.sh"
@@ -136,9 +136,11 @@ backup_old()
 
 rm_curent()
 {
-  rm -f "${FRONTEND}"
-  if [ "${GCC_BACKEND}" = yes ] ; then
-    rm -f "${BACKEND}"
+  if [ "${NOACTION}" != yes ] ; then
+    rm -f "${FRONTEND}"
+    if [ "${GCC_BACKEND}" = yes ] ; then
+      rm -f "${BACKEND}"
+    fi
   fi
 }
